# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Audio Streaming API
  version: 0.1.0
  description: Shared contract for the central backend and future local mini-backends.

paths:
  /audios:
    get:
      summary: List track metadata
      responses:
        '200':
          description: Track list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
    post:
      summary: Upload an audio file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                  description: Display name stored in the catalog (defaults to the file name without extension)
      responses:
        '201':
          description: File uploaded
        '400':
          description: Missing file or invalid name
        '409':
          description: A track with this name already exists

  /audios/{id}:
    get:
      summary: Stream an audio file
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Full audio stream
        '206':
          description: Partial content for range requests
        '404':
          description: Track not found
    delete:
      summary: Delete a track
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Track deleted
        '404':
          description: Track not found

  /sync/status:
    get:
      summary: Sync status between central and local catalogs
      responses:
        '200':
          description: Pending downloads and uploads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatus'

components:
  schemas:
    Track:
      type: object
      required: [id, path, name, extension]
      properties:
        id:
          type: integer
        path:
          type: string
        name:
          type: string
        extension:
          type: string
        contentHash:
          type: string
          description: SHA-256 content hash for duplicate detection
        checksum:
          type: string
          description: Reserved for future sync support
        updatedAt:
          type: string
          format: date-time
          description: Reserved for future sync support

    SyncStatus:
      type: object
      properties:
        pendingDownloads:
          type: array
          items:
            type: integer
        pendingUploads:
          type: array
          items:
            type: integer